refactor(bottom navigation): export bottom-navigation instead of tab bar
