+++ package.json
@@ -2,7 +2,7 @@
   "name": "taro-material",
   "private": true,
   "author": "lico",
-  "version": "0.6.0-8",
+  "version": "0.6.0-10",
   "description": "Mini Program components that implement Google's Material Design.",
   "keywords": [
     "react",
