+++ packages/taro-material/src/BottomNavigation/index2.js
@@ -1,0 +1,3 @@
+import BottomNavigation from './BottomNavigation';
+
+export default BottomNavigation;
