--- packages/taro-material/src/TabBar/index2.js
@@ -1,3 +1,0 @@
-import TabBar from './TabBar';
-
-export default TabBar;
