+++ packages/taro-material/src/index.js
@@ -66,6 +66,7 @@ export { default as RMAffix } from './Affix';
 export { default as RMAsyncSelect } from './AsyncSelect';
 export { default as RMAutocomplete } from './Autocomplete';
 export { default as RMBackTop } from './BackTop';
+export { default as RMBottomNavigation } from './BottomNavigation';
 export { default as RMButton } from './Button';
 export { default as RMCarousel } from './Carousel';
 export { default as RMCheckbox } from './Checkbox';
@@ -89,7 +90,6 @@ export { default as RMRadioGroup } from './RadioGroup';
 export { default as RMSearch } from './Search';
 export { default as RMSelect } from './Select';
 export { default as RMSwitch } from './Switch';
-export { default as RMTabBar } from './TabBar';
 export { default as RMTag } from './Tag';
 export { default as RMTextArea } from './TextArea';
 export { default as RMTextField } from './TextField';
