+++ pages/pages/material/bottom-navigation/index.scss
@@ -1,0 +1,5 @@
+@import '../../../styles/theme.scss';
+
+.root {
+  height: 150vh;
+}
