--- pages/pages/material/tab-bar/index.scss
@@ -1,5 +1,0 @@
-@import "../../../styles/theme.scss";
-
-.root{
-  height: 150vh
-}
